+++ common/core/desktop/src/km_kbp_keyboard_api.cpp
@@ -15,6 +15,7 @@
 #include "processor.hpp"
 #include "kmx/kmx_processevent.hpp"
 #include "mock/mock_processor.hpp"
+#include "rust/rust_mock_processor.hpp"
 
 using namespace km::kbp;
 
@@ -28,6 +29,9 @@ namespace
     }
     else if (kb_path.suffix() == ".mock") {
       return new mock_processor(kb_path);
+    }
+    else if (kb_path.suffix() == ".rust_mock") {
+      return new rust_mock_processor(kb_path);
     }
     else {
       return new null_processor();