+++ linux/keyman-config/keyman_config/fcitx_util.py
@@ -1,0 +1,13 @@
+import subprocess
+import dbus
+
+
+def is_fcitx_running():
+    for service in dbus.SessionBus().list_names():
+        if service == 'org.fcitx.Fcitx5':
+            return True
+    return False
+
+
+def restart_fcitx():
+    subprocess.run(['fcitx5', '-r', '-d'])